refactor : inner class 추출 및 like EntityListener 구현
