+++ src/main/java/com/example/demo/domain/board/service/service/LikeService.java
@@ -5,11 +5,6 @@ import com.example.demo.domain.board.Repository.LikeRepository;
 import com.example.demo.domain.board.domain.dto.response.BoardTitleInfoResponse;
 import com.example.demo.domain.board.domain.entity.Board;
 import com.example.demo.domain.board.domain.entity.Like;
-import com.example.demo.domain.notification.domain.entity.Notification;
-import com.example.demo.domain.notification.domain.entity.NotificationUser;
-import com.example.demo.domain.notification.domain.vo.NotificationType;
-import com.example.demo.domain.notification.repository.NotificationRepository;
-import com.example.demo.domain.notification.repository.NotificationUserRepository;
 import com.example.demo.domain.user.domain.User;
 import com.example.demo.domain.user.repository.UserRepository;
 import com.example.demo.global.base.dto.page.GlobalPageResponse;
@@ -17,17 +12,16 @@ import com.example.demo.global.base.exception.ErrorCode;
 import com.example.demo.global.base.exception.ServiceException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
-import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 @Component
 @RequiredArgsConstructor
 public class LikeService {
-    private final LikeNotificationService likeNotificationService;
     private final LikeRepository likeRepository;
     private final BoardRepository boardRepository;
     private final UserRepository userRepository;
+    private final LikeNotificationService likeNotificationService;
 
     @Transactional
     public void increaseLike(Long userId, Long boardId) {
@@ -67,32 +61,5 @@ public class LikeService {
         board.getLikes().remove(like);
         user.getLikes().remove(like);
         likeRepository.delete(like);
-        likeNotificationService.deleteLikeNotification(like);
-    }
-
-
-    @Component
-    @RequiredArgsConstructor
-    static class LikeNotificationService {
-        private final NotificationRepository notificationRepository;
-        private final NotificationUserRepository notificationUserRepository;
-
-        @Async
-        @Transactional
-        public void saveLikeNotification(Like like) {
-            Notification notification = Notification.fromLikeEntity(like, NotificationType.BOARD_LIKE);
-            if (!like.getBoard().getUser().getId().equals(like.getUser().getId())) {
-                User user = like.getBoard().getUser();
-                Notification saved = notificationRepository.save(notification);
-                notificationUserRepository.save(NotificationUser.from(saved, user));
-            }
-
-        }
-
-        @Async
-        @Transactional
-        public void deleteLikeNotification(Like like) {
-            notificationRepository.deleteByInvokerIdAndInvokerType(like.getId(), NotificationType.BOARD_LIKE);
-        }
     }
 }
